requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
Markdown==3.8
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0